Rename `shared` back to `vendor`
