+++ commands/mv.js
@@ -1,6 +1,6 @@
 import assert from '../modules/assert.js'
 import glob from '../modules/dat-glob.js'
-import isGlob from '../shared/is-glob-v4.0.1.js'
+import isGlob from '../vendor/is-glob-v4.0.1.js'
 import joinPath from '../modules/join-path.js'
 import parsePath from '../modules/dterm-parse-path.js'
