+++ index.js
@@ -4,7 +4,7 @@ import {selectHome, getEnv} from './modules/dterm-home.js'
 import control from './modules/dterm-controller.js'
 import glob from './modules/dat-glob.js'
 import getStream from './modules/dterm-stream.js'
-import isGlob from './shared/is-glob-v4.0.1.js'
+import isGlob from './vendor/is-glob-v4.0.1.js'
 import joinPath from './modules/join-path.js'
 import loadCommand from './modules/dterm-load-command.js'
 import parseCommand from './modules/parse-command.js'
