+++ modules/glob-parent.js
@@ -1,4 +1,4 @@
-import isGlob from '../shared/is-glob-v4.0.1.js'
+import isGlob from '../vendor/is-glob-v4.0.1.js'
 import joinPath from './join-path.js'
 
 export default function parent (glob) {
